requirements.txt
pandas==2.0.3
numpy<2
pyfirmata
pyserial